more cleanups and back out some cruft
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/bounded-pointers-branch@36239 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/cppfiles.c
@@ -216,7 +216,7 @@ open_include_file (pfile, filename)
 	{
 	  cpp_ice (pfile,
 		   "node for '%s' exists, open failed, error '%s', value %lx\n",
-		   filename, strerror (errno), nd->value);
+		   filename, strerror (errno), (unsigned long) nd->value);
 	  destroy_include_file_node (nd->value);
 	}
       splay_tree_insert (pfile->all_include_files,
